Methods updated. Shading problem driving me crazy ...

+++ libvectors/srcs/dot_vectors.c
@@ -6,7 +6,7 @@
 /*   By: rotrojan <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/01/28 10:29:07 by rotrojan          #+#    #+#             */
-/*   Updated: 2021/01/28 12:58:05 by rotrojan         ###   ########.fr       */
+/*   Updated: 2021/02/22 15:05:41 by rotrojan         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,5 +14,5 @@
 
 double	dot_vectors(t_vector vec1, t_vector vec2)
 {
-	return ((double)(vec1.x * vec2.x + vec1.y * vec2.y + vec1.z * vec2.z));
+	return (vec1.x * vec2.x + vec1.y * vec2.y + vec1.z * vec2.z);
 }
